Uploaded the Boxed-In Game created with teem for school project
contains the complete work of all teem members as of the end of the
class.  It still needs some work, currently there are only 4 levels
completed, the Wraith pathfinding needs to be fixed (it gets stuck on
walls),  the levels sizing needs to be enlarged (the player gets stuck
on corners),  the art assets need to be finished,  need to smooth out
mechanics and object interaction.